Define defaults in a separate file.
