+++ src/defaults.js
@@ -1,0 +1,22 @@
+const defaults = {
+  prefix: 'formPrefill',
+  storageKeys: function () {
+    return {
+      read: 'key',
+      write: 'key'
+    }
+  },
+  map: {},
+  exclude: '[data-form-prefill-exclude]',
+  include: '[data-form-prefill-include]',
+  stringPrefix: 's',
+  listPrefix: 'l',
+  stores: [],
+  useSessionStore: true,
+  useLocalStore: false,
+  useCookies: false,
+  cookieDomain: '',
+  cookieMaxAge: Infinity
+}
+
+export { defaults }
